Major refactor of the validation API.
It is now built with a very simple static entrypoint in the form of Missing.Validation.Validator (as before).
The news is that the static Validator now uses an instance of Missing.Validation.Internal.InternalValidator to perform
the actual validation.
Also, the "type scoring" (used to identify which ValidationSpecification to use) has been extracted into its own internal classes and
it no longer uses a custom comparer - it uses Missing.Collections.OrderedList instead.

Updated documentation.

The method "EachPrimitive<TItem>(..)" now performs a check to make sure that TItem is in fact a primitive type. It throws an
ArgumentException if it isn't.